requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.12.0=pypi_0
appnope=0.1.2=py38h50d1736_1
argon2-cffi=20.1.0=py38h96a0964_2
argparse=1.4.0=pypi_0
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
bottleneck=1.3.2=py38hf1fa96c_1
bzip2=1.0.8=h1de35cc_0
c-ares=1.17.2=h0d85af4_0
ca-certificates=2021.10.8=h033912b_0
cachetools=4.2.4=pypi_0
cairo=1.16.0=he43a7df_1008
certifi=2021.10.8=py38h50d1736_0
cffi=1.14.6=py38h9688ba1_0
charset-normalizer=2.0.9=pypi_0
click=8.0.1=pypi_0
cvxpy=1.1.15=py38h50d1736_0
cvxpy-base=1.1.15=py38ha53d530_0
cycler=0.10.0=pypi_0
cython=0.29.24=py38h23ab428_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.4=pypi_0
ecos=2.0.8=py38hbe852b5_1
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3.2=h4dad6da_0
flatbuffers=2.0=pypi_0
fontconfig=2.13.1=h10f422b_1005
freetype=2.10.4=ha233b18_0
future=0.18.2=py38h50d1736_3
gast=0.4.0=pypi_0
gettext=0.21.0=h7535e17_0
gitdb=4.0.7=pyhd8ed1ab_0
gitpython=3.1.23=pyhd8ed1ab_1
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.54.0=pypi_0
graphite2=1.3.13=h2e338ed_1001
grpcio=1.42.0=pypi_0
h5py=3.6.0=pypi_0
harfbuzz=3.0.0=h159f659_1
hdf5=1.12.1=nompi_h2f0ef1a_101
icu=68.1=h74dc148_0
idna=3.3=pypi_0
importlib-metadata=4.6.4=py38h50d1736_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.3.0=hecd8cb5_3375
ipykernel=5.5.5=py38h6c79ece_0
ipython=7.26.0=py38h5fd9f69_0
ipython_genutils=0.2.0=py_1
jasper=2.0.14=hee3b9ff_2
jbig=2.1=h0d85af4_2003
jedi=0.18.0=py38h50d1736_2
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pypi_0
jpeg=9d=hbcb3906_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=7.0.1=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h50d1736_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
krb5=1.19.2=hcfbf3a7_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=2.2.1=h046ec9c_0
libblas=3.9.0=11_osx64_mkl
libcblas=3.9.0=11_osx64_mkl
libclang=12.0.0=pypi_0
libcurl=7.79.1=hf45b732_1
libcxx=12.0.1=habf9029_0
libdeflate=1.7=h35c211d_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libglib=2.68.4=hd556434_0
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
liblapack=3.9.0=11_osx64_mkl
liblapacke=3.9.0=11_osx64_mkl
libnghttp2=1.43.0=h6f36284_1
libopencv=4.5.3=py38ha082d02_5
libpng=1.6.37=ha441bb4_0
libprotobuf=3.18.1=hcf210ce_0
libsodium=1.0.18=hbcb3906_1
libssh2=1.10.0=h52ee1ee_2
libtasn1=4.16.0=h9ed2024_0
libtiff=4.3.0=h1167814_1
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.2.1=h0d85af4_0
libxml2=2.9.12=h93ec3fd_0
libzlib=1.2.11=h9173be1_1013
llvm-openmp=12.0.1=hda6cdc1_1
lz4-c=1.9.3=h23ab428_1
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py38h96a0964_0
matplotlib=3.4.3=pypi_0
matplotlib-base=3.3.1=py38h181983e_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py38h96a0964_1004
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.0=py38h4a7008c_2
mkl_random=1.2.2=py38hb2f4e1b_0
nbclient=0.5.4=pyhd8ed1ab_0
nbconvert=6.1.0=py38h50d1736_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
nltk=3.6.2=pypi_0
notebook=6.4.3=pyha770c72_0
numexpr=2.7.3=py38h5873af2_1
numpy=1.21.2=pypi_0
numpy-base=1.20.3=py38he0bd621_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
opacus=0.14.0=pypi_0
opencv=4.5.3=py38h50d1736_5
openh264=2.1.1=hfd3ada9_0
openjpeg=2.3.0=hb95cd4c_1
openssl=1.1.1l=h0d85af4_0
opt-einsum=3.3.0=pypi_0
osqp=0.6.2=py38ha53d530_2
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.2=py38h5008ddb_0
pandoc=2.14.2=h0d85af4_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py38ha4cf6ea_0
pip=21.2.2=py38hecd8cb5_0
pixman=0.40.0=hbcb3906_0
prometheus_client=0.11.0=pyhd8ed1ab_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyha770c72_0
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
py-opencv=4.5.3=py38h5fd9f69_5
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.10.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h96a0964_2
python=3.8.5=h26836e1_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.9.1=cpu_py38h7f09a23_3
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.2.1=py38h23ab428_1
qdldl-python=0.1.5=py38ha53d530_1
readline=8.1=h9ed2024_0
regex=2021.8.21=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rouge=1.0.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=0.23.2=py38h959d312_0
scipy=1.7.1=py38hd329d04_0
scs=2.1.4=py38hd383344_0
seaborn=0.11.0=py_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h6db0672_2
smmap=3.0.5=pyh44b312d_0
sqlite=3.36.0=hce871da_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.11.1=py38h50d1736_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.9.0=py38h0a49c68_0_cpu
tornado=6.1=py38h96a0964_1
tqdm=4.62.2=pypi_0
traitlets=5.0.5=py_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
wrapt=1.13.3=pypi_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h9173be1_1013
zstd=1.5.0=h582d3a0_0
